# src/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: Upgrade Journey
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /tmf-api/agreementManagement/v4/agreement/{msisdn}:
    get:
      tags:
        - Plan Details
      summary: Current Contract summary
      operationId: getPlanDetails
      consumes:
        - application/json
      parameters:
        - in: path
          name: msisdn
          required: true
          type: string
          description: Msisdn of account.
      responses:
        200:
          description: Plan Details retrieved successfully.
          schema:
            $ref: "#/definitions/PlanDetails-Model"
        400:
          description: Missing required input parameter.
        404:
          description: No Data found.
        424:
          description: Upstream call failed with a 500. Response will be identical to upstream failure.
        500:
          description: Technical error occured, service failure.
...
  /tmf-api/usageManagement/v4/usage/consumptions:
    get:
      tags:
        - Consumption Usage
    
      description: Gets the customers consumption
      operationId: getConsumptionInfo
      parameters:
        - name: msisdn
          in: query
          type: string
          required: true
          description: The msisdn to retrieve the customers consumption
      responses:
        200:
          description: The customers consumption was successfully returned
          schema:
            $ref: "#/definitions/Consumption-Usage"
        "400":
          description: Missing required input parameter
        "424":
          description: Upstream call failed with a 500. Response will be identical to upstream failure
        "500":
          description: Technical error occured, service failure.        
...
definitions:
...
  PlanDetails-Model:
    type: object
    properties:
      agreementItem:
        type: array
        items:
          type: object
          properties:
            product:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                  name:
                     type: string
            termOrCondition:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                  validFor:
                    type: object
                    properties:
                      endDateTime:
                        type: string
                      startDateTime:
                        type: string
      characteristic:
        type: array
        items:
          type: object
          properties:
            name:
              type: string
            value:
              type: string
              
        # Get EligibilityCheck
...
  Consumption-Usage:
    type: object
    properties:
       usageCharacteristic:
            type: array
            items:
              type: string    
...